refactor: 학습하기 화면 컴포넌트 StudyScreen 파일로 이동
